+++ test/services/DummyCommandableCloudFunctionService_test.go
@@ -20,7 +20,7 @@ func newDummyCommandableCloudFunctionServiceTest() *DummyCommandableCloudFunctio
 func (c *DummyCommandableCloudFunctionServiceTest) setup(t *testing.T) {
 	config := cconf.NewConfigParamsFromTuples(
 		"logger.descriptor", "pip-services:logger:console:default:1.0",
-		"service.descriptor", "pip-services-dummies:service:commandable-gcp-function:default:1.0",
+		"service.descriptor", "pip-services-dummies:service:commandable-cloudfunc:default:1.0",
 	)
 
 	ctx := context.Background()